New copylog command
This command copies recovery log from /tmp to a defined destination.
In twrp it is the button, which copies recovery log from /tmp to /sdcard, with the difference that with 'copylog' command you can copy it to your favourite destination.
You can get more info about this command by running $ copylog --help